Staff Operations >Implement Conformation Messages During Process

+++ src/StaffOperations.cpp
@@ -29,8 +29,7 @@ void StaffOperations::addDVDsOfNewMovie(MovieCollection& movies) {
     std::string title = promptTitle();
     bool duplicate = movies.search(title);
     if (duplicate) {
-        cout << "\n\t\t\"" << formatString(title)
-             << "\" has already existed in the collection" << endl;
+        cout << "\n\t\t\"" << formatString(title)<< "\" has already existed in the collection" << endl;
         cout << "\n\t\tPlease add a new movie" << endl << endl;
         return;
     }
@@ -46,9 +45,13 @@ void StaffOperations::addDVDsOfNewMovie(MovieCollection& movies) {
     if (!duplicate) {
         Movie newMovie(title, starring, director, genre, classification,
         releaseDate, duration, numDVDs);
-        movies.insert(newMovie);
-        cout << "\n\t\t\"" << formatString(title)
-             << "\" has been added to the collection" << endl << endl;
+        cout<<"\n\t\tDo You Want To Add Movie\""<< formatString(title)<< "\" ?"<<endl;
+        if(displayConfirmation()==1){
+            movies.insert(newMovie);
+            cout << "\n\t\t\"" << formatString(title)<< "\" has been added to the collection" << endl << endl;
+        }
+        else
+            cout << "\n\t\tMovie Adding Operation Has Been Rejected.\n\t\tBack To Main Menu" << endl << endl;
     }
     cin.ignore(numeric_limits<streamsize>::max(), '\n');
 }
@@ -65,16 +68,18 @@ void StaffOperations::addDVDsOfExistingMovie(MovieCollection& movies) {
     }
     int numDVDs = promptNumDVDs();
     if (exist) {
-        Movie* movieAddress = movies.getItemAddress(title);
-        movieAddress->addDVDs(numDVDs);
-        cout << "\n\t\t" << numDVDs << " DVDs have been added to \""
-             << formatString(title) << "\"" << endl << endl;
-        cout << "\t\t" << setw(15) << left << "Title" << setw(20)
-             << right << "Number of DVDs" << endl << endl;
-        cout << "\t\t" << setw(15) << left << formatString(title)
-             << setw(20) << right << movieAddress->getNumDVDs() << endl << endl;
-        movieAddress = NULL;
-        delete movieAddress;
+        cout<<"\n\t\tDo You Want To Add (" << numDVDs << ") DVDs To \"" << formatString(title) <<"\""<<endl;
+        if(displayConfirmation()==1){
+            Movie* movieAddress = movies.getItemAddress(title);
+            movieAddress->addDVDs(numDVDs);
+            cout << "\n\t\t" << numDVDs << " DVDs have been added to \""<< formatString(title) << "\"" << endl << endl;
+            cout << "\t\t" << setw(15) << left << "Title" << setw(20)<< right << "Number of DVDs" << endl << endl;
+            cout << "\t\t" << setw(15) << left << formatString(title)<< setw(20) << right << movieAddress->getNumDVDs() << endl << endl;
+            movieAddress = NULL;
+            delete movieAddress;
+        }
+        else
+            cout << "\n\t\tMovie Adding DVD Operation Has Been Rejected.\n\t\tBack To Main Menu" << endl << endl;
     }
     cin.ignore(numeric_limits<streamsize>::max(), '\n');
 }
@@ -94,22 +99,24 @@ void StaffOperations::removeDVDs(MovieCollection& movies) {
         Movie* movieAddress = movies.getItemAddress(title);
         if (movieAddress->getRecord().empty() && movieAddress->getNumDVDs() == 0) {
             cout << "\n\t\tNumber of DVD is 0" << endl;
-        } else if (movieAddress->getRecord().empty() && movieAddress->getNumDVDs() != 0) {
-            if (movieAddress->getNumDVDs() - numDVDs < 0) {
-                cout << "\n\t\tThe number entered is greater the number of existing DVDs"
-                     << endl << endl;
-            } else {
-                movieAddress->removeDVDs(numDVDs);
-                cout << "\n\t\t" << numDVDs << " DVDs have been removed from \""
-                     << formatString(title) << "\"" << endl << endl;
-                cout << "\t\t" << setw(15) << left << "Title" << setw(20) << right
-                     << "Number of DVDs" << endl << endl;
-                cout << "\t\t" << setw(15) << left << formatString(title) << setw(20)
-                     << right << movieAddress->getNumDVDs() << endl << endl;
-            }
-        } else {
-        cout << "\n\t\tERROR: One or more DVDs are being rented." << endl << endl;
-        }
+        } else
+            if (movieAddress->getRecord().empty() && movieAddress->getNumDVDs() != 0) {
+                if (movieAddress->getNumDVDs() - numDVDs < 0) {
+                    cout << "\n\t\tThe number entered is greater the number of existing DVDs" << endl << endl;
+                    }
+                    else {
+                        cout<<"\n\t\tDo You Want To Remove (" << numDVDs << ") DVDs From \"" << formatString(title) <<"\""<<endl;
+                        if(displayConfirmation()==1){
+                            movieAddress->removeDVDs(numDVDs);
+                            cout << "\n\t\t" << numDVDs << " DVDs have been removed from \"" << formatString(title) << "\"" << endl << endl;
+                            cout << "\t\t" << setw(15) << left << "Title" << setw(20) << right << "Number of DVDs" << endl << endl;
+                            cout << "\t\t" << setw(15) << left << formatString(title) << setw(20)<< right << movieAddress->getNumDVDs() << endl << endl;
+                            }
+                        else
+                            cout << "\n\t\tMovie Remove DVD Operation Has Been Rejected.\n\t\tBack To Main Menu" << endl << endl;
+                        }
+                }else
+                    cout << "\n\t\tERROR: One or more DVDs are being rented." << endl << endl;
         movieAddress = NULL;
         delete movieAddress;
     }
@@ -126,17 +133,23 @@ void StaffOperations::removeMovie(MovieCollection& movies) {
         displayMovieNotFoundMessage();
         return;
     } else {
+
         Movie* movieAddress = movies.getItemAddress(title);
         if (movieAddress->getRecord().empty()) {
-            movies.erase(title);
-            cout << "\n\t\t\"" << formatString(title)
-                 << "\" has been remove from the collection" << endl << endl;
+            cout<<"\n\t\tDo You Want To Remove \"" << formatString(title) <<"\""<<endl;
+                if(displayConfirmation()==1){
+                    movies.erase(title);
+                    cout << "\n\t\t\"" << formatString(title)<< "\" has been remove from the collection" << endl << endl;
+                }
+                else
+                    cout << "\n\t\tMovie Remove Operation Has Been Rejected.\n\t\tBack To Main Menu" << endl << endl;
         } else {
             cout << "\n\t\tERROR: One or more DVDs are being rented." << endl << endl;
         }
         movieAddress = NULL;
         delete movieAddress;
     }
+    cin.ignore(numeric_limits<streamsize>::max(), '\n');
 }
 
 /// \brief Adds a new customer to the customer collection.
@@ -157,8 +170,14 @@ void StaffOperations::registerCustomer(CustomerCollection& customers) {
     cout << "\n\t\tCustomer's username: " << newCustomer.getUsername() << endl;
     std::string password = promptPassword();
     newCustomer.setPassword(password);
-    customers.addCustomer(newCustomer);
-    cout << "\n\t\t\"" << formatString(name) << "\" has been registered." << endl << endl;
+    cout<<"\n\t\tDo You Want To Register \""<<formatString(name)<<"\" ?"<<endl;
+    if(displayConfirmation() == 1){
+        customers.addCustomer(newCustomer);
+        cout << "\n\t\t\"" << formatString(name) << "\" has been registered." << endl << endl;
+    }
+    else
+        cout << "\n\t\tCustomer's Registration Operation has Rejected\n\t\tBack To Main Menu"<<endl;
+     cin.ignore(numeric_limits<streamsize>::max(), '\n');
 }
 
 /// \brief Removes an existing customer from the customer collection
@@ -181,8 +200,16 @@ void StaffOperations::removeCustomer(CustomerCollection& customers) {
             cout << "\n\t\tERROR: The customer is currently renting "
                  << customers.customerAt(index).getRecord().size() << " DVDs" << endl << endl;
         } else {
-            customers.deleteCustomer(name);
-            cout << "\n\t\t\"" << formatString(name) << "\" has been deleted" << endl << endl;
+            cout<<"\n\t\tDo You Want to Remove \""<<formatString(name)<<"\" From The List ?"<<endl;
+            if(displayConfirmation() == 1){
+                customers.deleteCustomer(name);
+                cout << "\n\t\t\"" << formatString(name) << "\" has been deleted" << endl << endl;
+            }
+            else{
+                cout << "\n\t\tRemove Customer's Operation has Rejected\n\t\tBack To Main Menu"<< endl << endl;
+
+            }
+            cin.ignore(numeric_limits<streamsize>::max(), '\n');
         }
     }
 }
@@ -493,5 +520,26 @@ void StaffOperations::displayMovieNotFoundMessage() {
 
 /// \brief Displays an error message when the customer collection does not contain the specified customer.
 void StaffOperations::displayCustomerNotFoundMessage() {
-    cout << "\n\t\Customer not found" << endl << endl;
+    cout << "\n\t\tCustomer not found" << endl << endl;
+}
+int StaffOperations::displayConfirmation(){
+    int confirm;
+    cout << "\t\t1) Yes" << endl;
+    cout << "\t\t2) No" << endl << endl;
+    do {
+        cout << "\t\tEnter an option: ";
+        cin >> confirm;
+        while (!cin) {
+            cin.clear();
+            cin.ignore(numeric_limits<streamsize>::max(), '\n');
+            displayErrorMessage();
+            cout << "\t\tEnter an option: ";
+            cin >> confirm;
+            cin.ignore(numeric_limits<streamsize>::max(), '\n');
+        }
+        if (confirm < 1 || confirm > 2) {
+            displayErrorMessage();
+        }
+    } while (confirm < 1 || confirm > 2);
+    return confirm;
 }
